Backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9
openpyxl
